move code into ./lib
